requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn==0.24.0
python-multipart==0.0.6

# Document Processing
python-docx==0.8.11
PyPDF2==3.0.1
azure-cognitiveservices-vision-computervision==0.9.0
azure-ai-formrecognizer==3.3.0
fuzzywuzzy

# NLP & ML
spacy==3.7.2
transformers==4.35.2
torch>=2.2.0
openai==1.3.5

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9

# Utilities
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.1.0
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0

# New packages
rapidfuzz
pymupdf 